Deep Learning Models - Training and Deployment Files/Flask Model APIs/im-image-app-main/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
gunicorn
tensorflow-cpu
Pillow
